backend/requirements.txt
charset-normalizer==3.3.2
Django==3.2.3
django-cors-headers==3.13.0
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
drf-extra-fields==3.7.0
flake8==6.0.0
gunicorn==20.1.0
pillow==10.4.0
psycopg2-binary==2.9.3
python-dotenv==1.0.1
requests==2.32.3
shortuuid==1.0.13
webcolors==1.11.1

